change weekly plot
