+++ datas/goal_states.csv
@@ -1,10 +1,10 @@
 ﻿date,total_usage_real,total_usage_goal,unlock_real,unlock_goal,app_usage_app,app_usage_real,app_usage_goal
-2023-05-02,411,570,196,70,Instagram,210,180
-2023-05-03,372,780,NA,NA,Youtube,176,150
-2023-05-04,581,240,184,80,Instagram,211,90
-2023-05-05,586,270,51,100,NA,NA,NA
-2023-05-06,492,570,101,130,Youtube,134,150
+2023-05-02,641,570,196,70,Instagram,210,180
+2023-05-03,853,780,NA,NA,Youtube,176,150
+2023-05-04,520,240,184,80,Instagram,211,90
+2023-05-05,486,270,51,100,NA,NA,NA
+2023-05-06,505,570,101,130,Youtube,134,150
 2023-05-07,NA,NA,62,200,Instagram,224,90
-2023-05-08,752,330,NA,NA,NA,NA,NA
-2023-05-09,432,840,135,170,Instagram,175,180
-2023-05-10,696,450,60,140,Instagram,183,210
+2023-05-08,504,330,NA,NA,NA,NA,NA
+2023-05-09,601,840,135,170,Instagram,175,180
+2023-05-10,604,450,60,140,Instagram,183,210
